Extract database methods for database/
