+++ backend/app/database/mysql.go
@@ -1,0 +1,39 @@
+package database
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/jmoiron/sqlx"
+)
+
+type MySQLConnectionEnv struct {
+	Host     string
+	Port     string
+	User     string
+	DBName   string
+	Password string
+}
+
+func NewMySQLConnectionEnv() *MySQLConnectionEnv {
+	return &MySQLConnectionEnv{
+		Host:     getEnv("MYSQL_HOST", "127.0.0.1"),
+		Port:     getEnv("MYSQL_PORT", "3306"),
+		User:     getEnv("MYSQL_USER", "ouchi"),
+		DBName:   getEnv("MYSQL_DBNAME", "ouchi"),
+		Password: getEnv("MYSQL_PASS", "ouchi"),
+	}
+}
+
+func getEnv(key, defaultValue string) string {
+	val := os.Getenv(key)
+	if val != "" {
+		return val
+	}
+	return defaultValue
+}
+
+func (mc *MySQLConnectionEnv) ConnectDB() (*sqlx.DB, error) {
+	dsn := fmt.Sprintf("%v:%v@tcp(%v:%v)/%v", mc.User, mc.Password, mc.Host, mc.Port, mc.DBName)
+	return sqlx.Open("mysql", dsn)
+}
